requirements.txt
selenium==4.16.0
pytest==7.4.0
PyHamcrest==2.1.0
allure-pytest==2.13.3